Started setting up test directories to actually ensure that leader election is working properly.

+++ src/java/SergeantCohort/CohortInfo.java
@@ -4,10 +4,10 @@ public class CohortInfo
 {
     public final String ip_addr_or_hostname;
     public final int port;
-    public final int cohort_id;
+    public final long cohort_id;
 
     public CohortInfo (
-        String ip_addr_or_hostname, int port, int cohort_id)
+        String ip_addr_or_hostname, int port, long cohort_id)
     {
         this.ip_addr_or_hostname = ip_addr_or_hostname;
         this.port = port;
